move constants for tooltip to own file, mirror style of button directory structure
