+++ assets/src/design-system/components/tooltip/constants.js
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export const TOOLTIP_POSITIONS = {
+  TOP_CENTER: 'top_center',
+  TOP_RIGHT: 'top_right',
+  TOP_LEFT: 'top_left',
+  BOTTOM_CENTER: 'bottom_center',
+  BOTTOM_RIGHT: 'bottom_right',
+  BOTTOM_LEFT: 'bottom_left',
+};
